[Update]投稿機能

+++ app/controllers/public/favorites_controller.rb
@@ -1,19 +1,17 @@
 class Public::FavoritesController < ApplicationController
-  
+
   before_action :authenticate_user!
 
   def create
-    post = Post.find(params[:post_id])
-    favorite = current_user.favorites.new(post_id: post.id)
-    favorite.save
-    redirect_to request.referer
+    @post = Post.find(params[:post_id])
+    current_user.favorites.find_or_create_by(post_id: @post.id)
+    render :favorite
   end
-  
+
   def destroy
-    post = Post.find(params[:post_id])
-    favorite = current_user.favorites.find_by(post_id: post.id)
-    favorite.destroy
-    redirect_to request.referer
+    @post = Post.find(params[:post_id])
+    favorite = current_user.favorites.find_by(post_id: @post.id)
+    favorite.destroy if favorite
+    render :favorite
   end
-  
 end
